rpa-agent-server/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-pptx==1.0.2
python-docx==1.1.2
openpyxl==3.1.5
playwright==1.47.0
httpx==0.27.2
python-multipart==0.0.9
aiofiles==24.1.0
pydantic==2.9.2
